add clone animator
잔상 효과를 위해 다른 오브젝트의 뼈 행렬을 클론할 수 있도록 기능을 추가하였다.
인스턴싱을 하고 싶지만 2024 바이트의 사이즈만 구조적 버퍼를 사용할 수 있기 때문에 8096 바이트의 사이즈를 사용하는 뼈 행렬을 구조적 버퍼에 넘겨줄 수 없다. 따라서 어쩔 수 없이 하나하나 오브젝트를 생성해야 한다.

+++ Client/X-Engine/src/Mesh.cpp
@@ -518,8 +518,8 @@ void ModelObjectMesh::CreateSphere(float radius, int sliceCount, int stackCount)
 		indices.push_back(baseIndex + i + 1);
 	}
 
-	mVertexCnt = positions.size();
-	mIndexCnt = indices.size();
+	mVertexCnt = (int)positions.size();
+	mIndexCnt = (int)indices.size();
 
 	D3DUtil::CreateVertexBufferResource(positions, mVertexUploadBuffer, mVertexBuffer);
 	D3DUtil::CreateVertexBufferResource(normals, mNormalUploadBuffer, mNormalBuffer);
